// project-frontend/src/main/frontend/src/app/generated/rest/project/model/registerRequest.ts
/**
 * WTM_UAM
 * API zum Verwalten User
 *
 * The version of the OpenAPI document: v1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RegisterRequest { 
    email?: string;
    password?: string;
    username?: string;
}